Move tower storage into its own module
