+++ core/src/consensus.rs
@@ -3,6 +3,7 @@ use {
         heaviest_subtree_fork_choice::HeaviestSubtreeForkChoice,
         latest_validator_votes_for_frozen_banks::LatestValidatorVotesForFrozenBanks,
         progress_map::{LockoutIntervals, ProgressMap},
+        tower_storage::{SavedTower, TowerStorage},
     },
     chrono::prelude::*,
     solana_ledger::{ancestor_iterator::AncestorIterator, blockstore::Blockstore, blockstore_db},
@@ -15,7 +16,7 @@ use {
         hash::Hash,
         instruction::Instruction,
         pubkey::Pubkey,
-        signature::{Keypair, Signature, Signer},
+        signature::Keypair,
         slot_history::{Check, SlotHistory},
     },
     solana_vote_program::{
@@ -25,13 +26,10 @@ use {
     std::{
         cmp::Ordering,
         collections::{HashMap, HashSet},
-        fs::{self, File},
-        io::BufReader,
         ops::{
             Bound::{Included, Unbounded},
             Deref,
         },
-        path::PathBuf,
     },
     thiserror::Error,
 };
@@ -107,7 +105,7 @@ pub(crate) struct ComputedBankState {
 #[frozen_abi(digest = "GMs1FxKteU7K4ZFRofMBqNhBpM4xkPVxfYod6R8DQmpT")]
 #[derive(Clone, Serialize, Deserialize, Debug, PartialEq, AbiExample)]
 pub struct Tower {
-    node_pubkey: Pubkey,
+    pub(crate) node_pubkey: Pubkey,
     threshold_depth: usize,
     threshold_size: f64,
     vote_state: VoteState,
@@ -1225,122 +1223,6 @@ impl TowerError {
     }
 }
 
-pub trait TowerStorage: Sync + Send {
-    fn load(&self, node_pubkey: &Pubkey) -> Result<SavedTower>;
-    fn store(&self, saved_tower: &SavedTower) -> Result<()>;
-}
-
-#[derive(Debug, Default, Clone, PartialEq)]
-pub struct NullTowerStorage {}
-
-impl TowerStorage for NullTowerStorage {
-    fn load(&self, _node_pubkey: &Pubkey) -> Result<SavedTower> {
-        Err(TowerError::WrongTower(
-            "NullTowerStorage has no storage".into(),
-        ))
-    }
-
-    fn store(&self, _saved_tower: &SavedTower) -> Result<()> {
-        Ok(())
-    }
-}
-
-#[derive(Debug, Default, Clone, PartialEq)]
-pub struct FileTowerStorage {
-    pub tower_path: PathBuf,
-}
-
-impl FileTowerStorage {
-    pub fn new(tower_path: PathBuf) -> Self {
-        Self { tower_path }
-    }
-
-    pub fn filename(&self, node_pubkey: &Pubkey) -> PathBuf {
-        self.tower_path
-            .join(format!("tower-{}", node_pubkey))
-            .with_extension("bin")
-    }
-}
-
-impl TowerStorage for FileTowerStorage {
-    fn load(&self, node_pubkey: &Pubkey) -> Result<SavedTower> {
-        let filename = self.filename(node_pubkey);
-        trace!("load {}", filename.display());
-
-        // Ensure to create parent dir here, because restore() precedes save() always
-        fs::create_dir_all(&filename.parent().unwrap())?;
-
-        let file = File::open(&filename)?;
-        let mut stream = BufReader::new(file);
-        bincode::deserialize_from(&mut stream).map_err(|e| e.into())
-    }
-
-    fn store(&self, saved_tower: &SavedTower) -> Result<()> {
-        let filename = self.filename(&saved_tower.node_pubkey);
-        trace!("store: {}", filename.display());
-        let new_filename = filename.with_extension("bin.new");
-
-        {
-            // overwrite anything if exists
-            let mut file = File::create(&new_filename)?;
-            bincode::serialize_into(&mut file, saved_tower)?;
-            // file.sync_all() hurts performance; pipeline sync-ing and submitting votes to the cluster!
-        }
-        fs::rename(&new_filename, &filename)?;
-        // self.path.parent().sync_all() hurts performance same as the above sync
-        Ok(())
-    }
-}
-
-#[frozen_abi(digest = "Gaxfwvx5MArn52mKZQgzHmDCyn5YfCuTHvp5Et3rFfpp")]
-#[derive(Default, Clone, Serialize, Deserialize, Debug, PartialEq, AbiExample)]
-pub struct SavedTower {
-    signature: Signature,
-    data: Vec<u8>,
-    #[serde(skip)]
-    node_pubkey: Pubkey,
-}
-
-impl SavedTower {
-    pub fn new<T: Signer>(tower: &Tower, keypair: &T) -> Result<Self> {
-        let node_pubkey = keypair.pubkey();
-        if tower.node_pubkey != node_pubkey {
-            return Err(TowerError::WrongTower(format!(
-                "node_pubkey is {:?} but found tower for {:?}",
-                node_pubkey, tower.node_pubkey
-            )));
-        }
-
-        let data = bincode::serialize(tower)?;
-        let signature = keypair.sign_message(&data);
-        Ok(Self {
-            signature,
-            data,
-            node_pubkey,
-        })
-    }
-
-    pub fn try_into_tower(self, node_pubkey: &Pubkey) -> Result<Tower> {
-        // This method assumes that `self` was just deserialized
-        assert_eq!(self.node_pubkey, Pubkey::default());
-
-        if !self.signature.verify(node_pubkey.as_ref(), &self.data) {
-            return Err(TowerError::InvalidSignature);
-        }
-        bincode::deserialize(&self.data)
-            .map_err(|e| e.into())
-            .and_then(|tower: Tower| {
-                if tower.node_pubkey != *node_pubkey {
-                    return Err(TowerError::WrongTower(format!(
-                        "node_pubkey is {:?} but found tower for {:?}",
-                        node_pubkey, tower.node_pubkey
-                    )));
-                }
-                Ok(tower)
-            })
-    }
-}
-
 // Given an untimely crash, tower may have roots that are not reflected in blockstore,
 // or the reverse of this.
 // That's because we don't impose any ordering guarantee or any kind of write barriers
@@ -1391,7 +1273,8 @@ pub mod test {
     use super::*;
     use crate::{
         fork_choice::ForkChoice, heaviest_subtree_fork_choice::SlotHashKey,
-        replay_stage::HeaviestForkFailures, vote_simulator::VoteSimulator,
+        replay_stage::HeaviestForkFailures, tower_storage::FileTowerStorage,
+        vote_simulator::VoteSimulator,
     };
     use solana_ledger::{blockstore::make_slot_entries, get_tmp_ledger_path};
     use solana_runtime::bank::Bank;
@@ -1408,6 +1291,7 @@ pub mod test {
         collections::HashMap,
         fs::{remove_file, OpenOptions},
         io::{Read, Seek, SeekFrom, Write},
+        path::PathBuf,
         sync::Arc,
     };
     use tempfile::TempDir;
