+++ core/src/lib.rs
@@ -53,6 +53,7 @@ pub mod sigverify_shreds;
 pub mod sigverify_stage;
 pub mod snapshot_packager_service;
 pub mod test_validator;
+pub mod tower_storage;
 pub mod tpu;
 pub mod tree_diff;
 pub mod tvu;
